aider: Modify loop in `populateMenu` to destructure both `panelId` and `tag` from `panels` object entries.
# Aider chat conversation:

USER: in `populateMenu` the `forEach` loop uses the key `panelId`. In that loop I need access to the `tag` key as well

ASSISTANT: To access both `panelId` and `tag` within the `forEach` loop in `populateMenu`, we need to modify the loop to iterate over the entries of the `panels` object. This way, we can destructure both the key and value.

Here is the *SEARCH/REPLACE* block:

app.js
```javascript
<<<<<<< SEARCH
    Object.keys(panels).forEach(tag, panelId => {
=======
    Object.entries(panels).forEach(([panelId, { tag }]) => {
>>>>>>> REPLACE
```